DMPG '15 S3 - Zen Garden
Zen gardens are intended to serve as an intimate imitation of nature as to aid in meditation about the true meaning of life. Therefore, Blake has decided to spend her summer building one in her backyard.

Each flower has a harmony value, and so she's planted NN flowers in a row as to ensure a peaceful garden. However, she's discovered that MM pairs of flowers, when planted beside each other, cause a decrease in the overall harmony of the garden.

The total harmony value of a Zen garden is given by the sum of all flower harmony values minus the sum of all disturbance values. As she wants her garden to have the maximum harmony, she can remove flowers as to eliminate the disturbance they cause.

What is the maximum overall harmony value that Blake can hope to get out of her garden?

Input Specification

The first line of input will consist of the single integer NN (4≤N≤100000)(4≤N≤100000).

The second line will consist of NN space-separated integers representing the meditation values (between 11 and 100100) of flowers 11 to NN as they are planted in Blake's garden.

The next line will consist of the single integer MM (1≤M≤N2)(1≤M≤N2).

The next MM lines will each consist of two space-separated integers, aa (1≤a<N)(1≤a<N) and dd, (1≤d≤1000)(1≤d≤1000), representing that flower aa and a+1a+1 together cause a decrease in the overall harmony by dd. Removing either aa or a+1a+1 will remove the disturbance dd.

It is guaranteed that all flowers will at most be part of one pair.

Output Specification

Output the maximum possible overall meditation value of the garden.