CCC '00 J2 - 9966
Canadian Computing Competition: 2000 Stage 1, Junior #2

The digits 00, 11, and 88 look much the same if rotated 180180 degrees on the page (turned upside down). Also, the digit 66 looks much like a 99, and vice versa, when rotated 180180 degrees on the page. A multi-digit number may also look like itself when rotated on the page; for example 99669966 and 1080110801 do, but 999999 and 12341234 do not.

You are to write a program to count how many numbers from a given interval look like themselves when rotated 180180 degrees on the page. For example, in the interval [1…100][1…100] there are six: 11, 88, 1111, 6969, 8888, and 9696.

Your program should take as input two integers, mm and nn, which define the interval to be checked, 1≤m≤n≤320001≤m≤n≤32000. The output from your program is the number of rotatable numbers in the interval.

You may assume that all input is valid.